# Optimize WriteRTP