--- a/go.mod
+++ b/go.mod
@@ -2,7 +2,7 @@ module github.com/pions/srtp
 
 require (
 	github.com/pions/rtcp v1.1.0
-	github.com/pions/rtp v1.0.1
+	github.com/pions/rtp v1.1.0
 	github.com/pkg/errors v0.8.1
 	github.com/stretchr/testify v1.3.0
 )