The old EncryptRTP API causes an extra header Unmarshal and payload
Marshal. By passing the `*rtp.Packet` through the stack, we're able to
avoid both of these.

There was also an extra reallocation caused by using append.

```
name                 old time/op  new time/op  delta
EncryptRTP-8         1.94µs ± 3%  1.94µs ± 3%    ~
EncryptRTPInPlace-8  1.89µs ± 3%  1.89µs ± 4%    ~
DecryptRTP-8         1.80µs ± 3%  1.83µs ± 7%    ~
Write-8              1.95µs ± 8%  1.93µs ± 1%   ~
WriteRTP-8           2.11µs ± 6%  2.01µs ± 8%  -4.71%
```